+++ Sources/ZcashLightClientKit/Synchronizer.swift
@@ -175,7 +175,7 @@ public protocol Synchronizer {
     func latestHeight() throws -> BlockHeight
 
     /// Returns the latests UTXOs for the given address from the specified height on
-    func refreshUTXOs(address: String, from height: BlockHeight, result: @escaping (Result<RefreshedUTXOs, Error>) -> Void)
+    func refreshUTXOs(address: String, from height: BlockHeight) async throws -> RefreshedUTXOs
 
     /// Returns the last stored transparent balance
     func getTransparentBalance(accountIndex: Int) throws -> WalletBalance